debbugging fixes to TransaccionComprobanteUnlinkRequest

+++ src/OrionERP.Application/Features/Contabilidad/Transacciones/TransaccionComprobanteUnlinkRequest.cs
@@ -4,5 +4,5 @@ public sealed class TransaccionComprobanteUnlinkRequest
 {
   public int CurrentTransaccionId { get; set; }
   public int TempTransaccionId { get; set; }
-  public int ComprobanteId { get; set; }
+  public long ComprobanteId { get; set; }
 }
